requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiohttp-devtools==0.13.1
aiohttp-jinja2==1.2.0
aiohttp-session[secure]==2.9.0
aiohttp==3.6.2
aioredis==1.3.1
async-timeout==3.0.1      # via aiohttp, aioredis
attrs==19.3.0             # via aiohttp
cffi==1.13.2              # via cryptography
chardet==3.0.4            # via aiohttp
click==7.0                # via aiohttp-devtools
cryptography==2.8         # via aiohttp-session
devtools==0.5.1           # via aiohttp-devtools
hiredis==1.0.1            # via aioredis
idna==2.8                 # via yarl
jinja2==2.11.1            # via aiohttp-jinja2
markupsafe==1.1.1         # via jinja2
multidict==4.7.4          # via aiohttp, yarl
pycparser==2.19           # via cffi
pygments==2.5.2           # via aiohttp-devtools
python-decouple==3.3
six==1.14.0               # via cryptography
watchgod==0.5             # via aiohttp-devtools
yarl==1.4.2               # via aiohttp
